backend/requirements.txt
fastapi
pydantic
uvicorn
requests
google-genai